Deps: Update kotlin to 1.9.25
Release Notes: https://github.com/JetBrains/kotlin/releases/tag/v1.9.25

FYI: This is done because 'automattic-tracks' itself is on Kotlin
version '1.9.24', which makes the current Kotlin version for this
project incompatible, see build failure below:

"Class 'kotlin.Unit' was compiled with an incompatible version of
Kotlin. The binary version of its metadata is 1.9.0, expected version
is 1.6.0."

PS: I chose '1.9.25' instead of '1.9.24' because that version is the
latest Kotlin '1.9' version before Kotlin '2.0'.