Deps: Update dagger to 2.52
Release Notes: https://github.com/google/dagger/releases/tag/dagger-2.52

FYI: This is done because of the Kotlin '1.9.25' update, which makes it
incompatible with the current '2.38.1' and/or '2.40.1' version of
Dagger, see build failure below:

"Unsupported metadata version. Check that your Kotlin version is
>= 1.0: java.lang.IllegalStateException: Unsupported metadata version.
Check that your Kotlin version is >= 1.0"

PS: I chose '2.52' instead of the latest '2.57' because that version
is the latest Dagger version compatible with Kotlin '1.9'.