Deps: Update androidx lifecycle to 2.5.1
Release Notes: https://developer.android.com/jetpack/androidx/releases/
lifecycle#2.5.1

FYI: This is done because of the Dagger '2.52' update, which makes it
incompatible with the current '2.3.1' version of AndroidX Lifecycle, see
build failure below:

"Duplicate class androidx.lifecycle.ViewModelLazy/
androidx.lifecycle.ViewTreeViewModelKt found in modules 2.3.1 and 2.5.1"

PS: I chose '2.5.1' instead of the latest '2.9.2' because that version
is compatible with Kotlin '1.9', but also because I didn't want to
introduce more (unnecessary) changes to this project. For example, I
could have chosen '2.8.7' as well, which also is fully compatible with
Kotlin '1.9', but I could have chosen the latest '2.9.2' because this
version is only compatible with Kotlin '2+'.