Fix: Add an default simplestyle splash them point to simplestyle
FYI: This change is done because after all the updates, the app started
showing a blue overlay with an icon on every screen. With this fix and
for Android 12 devices and above (API 31) we are no longer depending on
this custom splash screen theme, but rather let the system default to
whichever is the default implementation of splashscreen, without us
adding any additional configuration on top of it and thus defaulting the
main application theme to 'Theme.Simplestyle' for Android 12 devices and
above.